Moved wait_and_fetch_net_data from common.
The function is needed in the data_fetch module, and the circular import
would prevent the module to access the common module, where essential
functions s.a. _ are located.
